backend/requirements.txt
fastapi
uvicorn
sqlmodel
apscheduler
openai
pypdf
python-multipart
google-generativeai
python-dotenv
reportlab
